UI tweaks, clean and optimizations
GREAT clean in code, and UI tweaks for better rendering. Modified the FrontController, so it's easier to invoke new Stage, change UC in Stage (see code for examples of using)

Added ComboBox to choose type of search
Added binding in home page for current user
Serialisation is still to be done but i added CategoryFactory, which blocks category to be mutable as constructor is protected so XMLEncoder can't access it. Try to find a way to do serialisation with maybe XMLCategory and XMLGitHubRepository (which implements Follow and GitHubRepository extends Repository) as proxy implementation of a composite pattern (I tried it, I think this can be done, but not really simple)

Several bug fixes as well